requirements.txt
pandas
matplotlib
wordcloud
telethon
cryptg
emoji
pytz
networkx
scipy
guppy3
tqdm